Added doc and moved class to new file
