+++ Backpack-SwiftUI/Calendar/Classes/Core/CalendarContainer.swift
@@ -153,48 +153,3 @@ struct CalendarContainer_Previews: PreviewProvider {
         }
     }
 }
-
-// MARK: - Item Visibility Detection in Scroll View
-private struct ItemVisibilityPreferenceKey: PreferenceKey {
-    static var defaultValue: [Int: CGRect] = [:]
-
-    static func reduce(value: inout [Int: CGRect], nextValue: () -> [Int: CGRect]) {
-        value.merge(nextValue(), uniquingKeysWith: { $1 })
-    }
-}
-
-private class ItemVisibilityObserver: ObservableObject {
-    @Published var visibleItems: [Int] = []
-
-    private let preferencePublisher = PassthroughSubject<[Int: CGRect], Never>()
-    private var cancellables = Set<AnyCancellable>()
-
-    private var lastEmittedVisibleItems: [Int] = []
-
-    init(parentProxy: GeometryProxy) {
-        preferencePublisher
-            .map { preferences in
-                let parentFrame = parentProxy.frame(in: .global)
-                return preferences.filter { (_, frame) in
-                    frame.intersects(parentFrame)
-                }.map { $0.key }
-            }
-            .removeDuplicates() // Ensures we don't debounce if the list hasn't changed
-            .debounce(for: .milliseconds(200), scheduler: DispatchQueue.main)
-            .sink { [weak self] newVisibleItems in
-                guard let self = self else { return }
-
-                // Emit only if there's a real change in visible items
-                if self.lastEmittedVisibleItems != newVisibleItems {
-                    self.visibleItems = newVisibleItems
-                    self.lastEmittedVisibleItems = newVisibleItems
-                }
-            }
-            .store(in: &cancellables)
-    }
-
-    func updatePreferences(_ preferences: [Int: CGRect]) {
-        preferencePublisher.send(preferences)
-    }
-
-}
